seperated main.js into main.js and RecipeList.js
