+++ src/recipe_list/recipe_list.html
@@ -7,7 +7,7 @@
   <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
   <link rel="stylesheet" type="text/css" href="./recipe_list.css">
   <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-  <script src="../scripts/main.js" type="module"></script>
+  <script src="../scripts/RecipeList.js" type="module"></script>
   <script src="../scripts/RecipeCard.js" type="module"></script>
   <script src="../scripts/sideBar.js"></script>
   <title>NaN Bread Recipes</title>
@@ -21,7 +21,7 @@
 <body>
     <header>
       <nav>
-        <a href="../src/home.html">
+        <a href="/src/home/home.html">
           <h1 class="title">oui oui baguette</h1>
         </a>
       </nav>
@@ -35,8 +35,8 @@
       <button class="openbtn" onclick="openNav()">&#9776;</button>
     </div>
     <div id="sideNavbar" class="sidebar">
-      <a href="../src/home.html">Home</a>
-      <a href="../src/recipe_list/recipe_list.html">Recipe List</a>
+      <a href="/src/home/home.html">Home</a>
+      <a href="/src/recipe_list/recipe_list.html">Recipe List</a>
       <a href="#Favorites">Favorites</a>
     </div>
     <main>
